pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "context-aware-env"
version = "1.0.0"
description = "A social-awareness RL environment that evaluates whether an LLM agent knows when to stay silent vs. proactively help."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "websockets>=12.0",
    "httpx>=0.25.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
include = ["context_aware_env*", "server*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.txt", "*.md"]
